Creating adapter for estate list
Every basic function for the adapter have been implemented, as well as
a ViewModel and a ViewHolder for these estate items.

The PropertiesListFragment also changed, as we now need to provide a
list a properties in the newInstance() function.

Note that MainViewPagerAdapter may not compile, as we did not provide
any list of properties to PropertiesListFragment.